fix newdashboard route
