+++ resources/views/SuperAdmin/adminsidebar.blade.php
@@ -13,7 +13,7 @@
     <div class="collapse navbar-collapse  w-auto " id="sidenav-collapse-main">
         <ul class="navbar-nav">
             <li class="nav-item">
-                <a class="nav-link {{ Request::is('newdashboard') ? 'active' : '' }}" href="/newdashboard">
+                <a class="nav-link {{ Request::is('dashboard') ? 'active' : '' }}" href="/dashboard">
                     <div
                         class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
                         <i class="ni ni-tv-2 text-primary text-sm opacity-10"></i>
@@ -22,7 +22,7 @@
                 </a>
             </li>
             <li class="nav-item">
-                <a class="nav-link {{ Request::is('newdashboard/create*') ? 'active' : '' }} " href="/newdashboard/create">
+                <a class="nav-link {{ Request::is('dashboard/create*') ? 'active' : '' }} " href="/dashboard/create">
                     <div
                         class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
                         <i class="ni ni-single-02 text-dark text-sm opacity-10"></i>
@@ -32,7 +32,7 @@
             </li>
             <li class="nav-item">
                 <a class="nav-link {{
-                    Request::is('newdashboard/show') ? 'active' : '' }} " href="/newdashboard/show">
+                    Request::is('dashboard/show') ? 'active' : '' }} " href="/dashboard/show">
                     <div
                         class="icon icon-shape icon-sm border-radius-md text-center me-2 d-flex align-items-center justify-content-center">
                         <i class="ni ni-calendar-grid-58 text-warning text-sm opacity-10"></i>
